requirements.txt
spotipy
python-decouple
bilibili_api